requirements.txt
aiogram==2.25.1
python-telegram-bot==29.1
requests==2.28.1
openai==0.27.8
python-dotenv==1.0.0
Flask==2.2.3
Werkzeug==2.2.3
gunicorn==20.1.0
